updated maps, marriage, etc., added wander, stable but buggy

+++ modules/initializemodel/initializemodel.go
@@ -18,7 +18,7 @@ func InitializeModel(configRoot string) (*types.Model, error) {
 		DeathRisk:      make(map[int]float64),
 		CumulativeProb: make(map[int]float64),
 		FreeParameters: make(map[string]int),
-		Map:            make(map[float64]map[float64]int),
+		Map:            make(map[int]map[int]int),
 	}
 
 	// Attempt to load each config file. Failure will be fatal.
